% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.r
\name{calcEnrichment}
\alias{calcEnrichment}
\title{Perform metabolite set enrichment for bmetenrichr enrichment objects}
\usage{
calcEnrichment(object, ...)
}
\arguments{
\item{object}{A bmetenrichr object.}

\item{n}{A integeter describing the number of bootstraps (default = 50).}
}
\value{
An object of class bmetenrich.
}
\description{
calcEnrichment() ranks metabolites of bmetenrichr object to perform bootstrapping metabolite set enrichment analysis
}
\examples{
myTestRun <- calcEnrichment(object = myTestRun, n = 100)


}
