% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.r
\name{enrichmentTable}
\alias{enrichmentTable}
\title{Export bootstrap enrichment analysis}
\usage{
enrichmentTable(object, ...)
}
\arguments{
\item{object}{A bmetenrichr object after enrichment analysis.}

\item{min.annotations}{An integer describing the minimal number of annotations each term should include (default = 2).}

\item{q.value.cutoff}{A numeric between 0 and 1. Only terms with q-values lower than this value will be displayed (default = 0.5).}
}
\value{
A data.frame
}
\description{
Export bootstrap enrichment analysis
}
\examples{

enrichmentTable(myTestRun)

}
