% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.r
\name{initLIONenrichment}
\alias{initLIONenrichment}
\title{Generate bmetenrichr enrichment object}
\usage{
initLIONenrichment(
  scmatrix,
  annotations,
  annotation.weights = NULL,
  conditions,
  include = NULL,
  pathway = "LION",
  termsOfInterest = "selection",
  condition.x = NULL,
  condition.y = NULL,
  ranking.by = "t.test"
)
}
\arguments{
\item{scmatrix}{A numeric matrix of n metabolites (rows) and m cells or measurments (columns).}

\item{annotations}{A list of length n, each element contains a vector of isomer names}

\item{annotation.weights}{A list of length n, each element contains a vector of isomer names}

\item{conditions}{A vector of length m with condition identifiers}

\item{include}{A logical vector of length n}

\item{condition.x}{first condition identifier for pairwise comparison}

\item{condition.y}{second condition identifier for pairwise comparison}
}
\value{
An object of class LIONenrichment.
}
\description{
Generate bmetenrichr enrichment object
}
\examples{
add(1, 1)
add(10, 1)
myTestRun <-
initLIONenrichment(scmatrix = scMatrix, 
                   annotations = my_annotations, 
                   annotation.weights = my_weights,
                   conditions = my_conditions,
                   condition.x = "A",
                   condition.y = "B" )


}
