% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.r
\name{plotEnrichment.bmetenrich}
\alias{plotEnrichment.bmetenrich}
\title{Plot bootstrap enrichment analysis}
\usage{
\method{plotEnrichment}{bmetenrich}(
  object,
  min.annotations = 2,
  q.value.cutoff = 0.1,
  plotIDs = FALSE,
  by.statistic = "ES"
)
}
\arguments{
\item{object}{A bmetenrichr object after enrichment analysis.}

\item{min.annotations}{An integer describing the minimal number of annotations each term should include}

\item{q.value.cutoff}{A numeric between 0 and 1. Only terms with q-values lower than this value will be displayed.}

\item{plotIDs}{A logical indicating whether term IDs should be displayed.}

\item{by.statistic}{A character indicating how the x-axis will be arranged. Can be either 'ES' (enrichment score) or 'q.value'.}
}
\value{
A ggplot2 object.
}
\description{
Plot bootstrap enrichment analysis
}
\examples{

plotEnrichment(myTestRun)

}
