% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.r
\name{rankScore}
\alias{rankScore}
\title{Rank metabolites for bmetenrichr enrichment object}
\usage{
rankScore(object, ...)
}
\arguments{
\item{object}{A bmetenrichr object.}

\item{ranking.by}{A character of either 't.test' (default) or 'wilcox.test', to rank metabolites using the respective statistic.}
}
\value{
An object of class bmetenrich.
}
\description{
rankScore() ranks metabolites of bmetenrichr object to perform bootstrapping metabolite set enrichment analysis
}
\examples{
myTestRun <-
rankScore(object = myTestRun, ranking.by = 't.test')


}
